+++ src/postgres-util/Cargo.toml
@@ -10,8 +10,6 @@ publish = false
 anyhow = "1.0.55"
 
 mz-ore = { path = "../ore", features = ["task"] }
-mz-repr = { path = "../repr" }
-mz-pgrepr = { path = "../pgrepr" }
 openssl = { version = "0.10.38", features = ["vendored"] }
 postgres-openssl = { git = "https://github.com/MaterializeInc/rust-postgres", branch = "mz-0.7.2" }
 tokio = { version = "1.17.0", features = ["fs"] }
