+++ test/pg-cdc/create-table-after-source.td
@@ -35,3 +35,4 @@ INSERT INTO t1 VALUES (1);
 
 # To keep this test passing we expect no result here, since that is now what we get
 > SELECT * FROM t1;
+1
